SecretTemplate watches for Secret InputResources
This introduces a watch mechanism for SecretTemplate when no service
account is specified, so for this case we can remove the periodic
re-sync - improving performance.

We still need a periodic resync for non-secret resources because we
use a user configurable service account for these resources.
Therefore we cannot use a single watch or shared informer and setting
up watches per secret template service account per GVR is too expensive.

Implementation notes:
- As mentioned only support watch when Service Account is not specified
- Rely on an internal tracker that can be referred when a secret event
in generated. This tracker allows a SecretTemplate to tracker N Secrets.
And for a given Secret, allows you to receive all interested
SecretTemplates.
- When a SecretTemplate is deleted we use 404 to clear the tracking
information. Even if we don't receive a reconcile immediately, if an
input resource changes, we will receive a notfication for the missing
SecretTemplate and then clear tracking that way.