Introduce SecretTemplate API (#66)
* Add basic SecretTemplate API
* Add custom statuses to SecretTemplate.

They are initially set to Unknown and updated to True/False
depending on the reconcile.  The Ready condition is an and of the
previous true/false conditions.
They are:
- InputResourcesFound => could read the .spec.inputResources
- TemplatingSucceeded => could evaluate json paths in inputresources
- SecretCreated => could create the child secret

* Add in a periodic sync of 30s.
* Type is immutable. So this implementation mimics the rest of
secretgen controller, which is to not update the type. We should revisit
this.

* SecretTemplate no longer cleans up secret on failure

Previously we deleted the secret if any error occurred during
reconcile. This was to guard against keeping data around that may have
been removed. However this previous behaviour also may lead to jittery secrets that
get deleted/recreated and that may bring down applications that are
mounting the secrets.

More context here https://github.com/vmware-tanzu/carvel-secretgen-controller/pull/66#discussion_r878107046

Co-authored-by: Glen Rodgers <grodgers@vmware.com>